ページ切り替え機能

+++ app/controllers/favorites_controller.rb
@@ -2,21 +2,29 @@ class FavoritesController < ApplicationController
   before_action :set_practice, only: [ :create, :destroy] 
 
   def create
-    favorite = current_user.favorites.build(practice_id: params[:practice_id])
-    favorite.save
-    render partial: "favorites/favorite", locals: { practice: @practice }
-    # respond_to do |format| 
-    #   format.js
-    # end
+    if user_signed_in?
+      favorite = current_user.favorites.build(practice_id: params[:practice_id])
+      favorite.save
+      render partial: "favorites/favorite", locals: { practice: @practice }
+      # respond_to do |format| 
+      #   format.js
+      # end
+    else
+      redirect_to new_user_session_path
+    end    
   end
 
   def destroy
-    favorite = Favorite.find_by(practice_id: params[:practice_id], user_id: current_user.id)
-    favorite.destroy
-    render partial: "favorites/favorite", locals: { practice: @practice }
-    # respond_to do |format|
-    #   format.js
-    # end
+    if user_signed_in?
+      favorite = Favorite.find_by(practice_id: params[:practice_id], user_id: current_user.id)
+      favorite.destroy
+      render partial: "favorites/favorite", locals: { practice: @practice }
+      # respond_to do |format|
+      #   format.js
+      # end
+    else
+      redirect_to new_user_session_path
+    end  
   end
 
   def set_practice
